+++ servidor.c
@@ -1,3 +1,5 @@
+#include "servidor.h"
+
 #include <arpa/inet.h>
 #include <netdb.h>
 #include <netinet/in.h>
@@ -19,46 +21,14 @@
 #define PACKET_TYPE_REQ 2
 #define PACKET_TYPE_REQ_ACK 4
 
-struct requisicao
-{
-    uint32_t value;
-};
-
-struct requisicao_ack
-{
-    uint32_t seqn;
-    uint32_t num_reqs;
-    uint64_t total_sum;
-};
-
-typedef struct __packet
-{
-    uint16_t type;
-    uint32_t seqn;
-    union
-    {
-        struct requisicao req;
-        struct requisicao_ack ack;
-    } data;
-} packet;
-
-typedef struct
-{
-    struct sockaddr_in addr;
-    uint32_t last_req;
-    uint64_t last_sum;
-} client_entry;
-
 client_entry clients[MAX_CLIENTS];
 int client_count = 0;
 uint32_t total_reqs = 0;
 uint64_t total_sum = 0;
 pthread_mutex_t lock = PTHREAD_MUTEX_INITIALIZER;
 
-int find_or_add_client(struct sockaddr_in *addr)
-{
-    for (int i = 0; i < client_count; i++)
-    {
+int find_or_add_client(struct sockaddr_in *addr) {
+    for (int i = 0; i < client_count; i++) {
         if (clients[i].addr.sin_addr.s_addr == addr->sin_addr.s_addr)
             return i;
     }
@@ -68,8 +38,7 @@ int find_or_add_client(struct sockaddr_in *addr)
     return client_count++;
 }
 
-void *handle_request(void *arg)
-{
+void *handle_request(void *arg) {
     struct
     {
         packet pkt;
@@ -83,8 +52,7 @@ void *handle_request(void *arg)
     int idx = find_or_add_client(&ctx->addr);
 
     pthread_mutex_lock(&lock);
-    if (pkt->seqn == clients[idx].last_req + 1)
-    {
+    if (pkt->seqn == clients[idx].last_req + 1) {
         total_reqs++;
         total_sum += pkt->data.req.value;
         clients[idx].last_req = pkt->seqn;
@@ -93,9 +61,7 @@ void *handle_request(void *arg)
         printf("%s client %s id_req %u value %u num_reqs %u total_sum %lu\n",
                timebuf, inet_ntoa(ctx->addr.sin_addr), pkt->seqn,
                pkt->data.req.value, total_reqs, total_sum);
-    }
-    else
-    {
+    } else {
         current_time(timebuf, sizeof(timebuf));
         printf("%s client %s DUP!! id_req %u value %u num_reqs %u total_sum %lu\n",
                timebuf, inet_ntoa(ctx->addr.sin_addr), pkt->seqn,
@@ -116,10 +82,8 @@ void *handle_request(void *arg)
     return NULL;
 }
 
-int main(int argc, char *argv[])
-{
-    if (argc != 2)
-    {
+int main(int argc, char *argv[]) {
+    if (argc != 2) {
         fprintf(stderr, "Uso: %s <porta>\n", argv[0]);
         exit(EXIT_FAILURE);
     }
@@ -129,8 +93,7 @@ int main(int argc, char *argv[])
 
     // Cria o socket com domínio AF_INET e tipo SOCK_DGRAM (protocolo UDP)
     int sock;
-    if ((sock = socket(AF_INET, SOCK_DGRAM, 0)) == -1)
-    {
+    if ((sock = socket(AF_INET, SOCK_DGRAM, 0)) == -1) {
         perror("Cannot ceate socket");
         exit(EXIT_FAILURE);
     }
@@ -138,14 +101,13 @@ int main(int argc, char *argv[])
     // Cria o container genérico que permite que o SO identifique a família do endereço
     struct sockaddr_in servaddr;
     memset((char *)&servaddr, 0, sizeof(servaddr));
-    servaddr.sin_family = AF_INET;                // IPv4
-    servaddr.sin_port = htons(port);              // Porta do servidor
-    servaddr.sin_addr.s_addr = htonl(INADDR_ANY); // Aceita qualquer endereço
+    servaddr.sin_family = AF_INET;                 // IPv4
+    servaddr.sin_port = htons(port);               // Porta do servidor
+    servaddr.sin_addr.s_addr = htonl(INADDR_ANY);  // Aceita qualquer endereço
     bzero(&(servaddr.sin_zero), 8);
 
     // Vincular o socket à porta
-    if (bind(sock, (struct sockaddr *)&servaddr, sizeof(servaddr)) < 0)
-    {
+    if (bind(sock, (struct sockaddr *)&servaddr, sizeof(servaddr)) < 0) {
         perror("Cannot bind socket");
         close(sock);
         exit(EXIT_FAILURE);
@@ -155,20 +117,16 @@ int main(int argc, char *argv[])
     current_time(timebuf, sizeof(timebuf));
     printf("%s num_reqs 0 total_sum 0\n", timebuf);
 
-    while (1)
-    {
+    while (1) {
         packet pkt;
         struct sockaddr_in cliaddr;
         socklen_t len = sizeof(cliaddr);
 
         recvfrom(sock, &pkt, sizeof(pkt), 0, (struct sockaddr *)&cliaddr, &len);
 
-        if (pkt.type == PACKET_TYPE_DISCOVERY)
-        {
+        if (pkt.type == PACKET_TYPE_DISCOVERY) {
             discovery_handle_request(sock, &cliaddr, len, (discovery_packet *)&pkt);
-        }
-        else if (pkt.type == PACKET_TYPE_REQ)
-        {
+        } else if (pkt.type == PACKET_TYPE_REQ) {
             pthread_t tid;
             void *ctx = malloc(sizeof(packet) + sizeof(cliaddr) + sizeof(socklen_t) + sizeof(int));
             memcpy(ctx, &(struct { packet pkt; struct sockaddr_in addr; socklen_t addrlen; int sock; }){pkt, cliaddr, len, sock}, sizeof(packet) + sizeof(cliaddr) + sizeof(socklen_t) + sizeof(int));
